@@ -1,1 +1,1 @@
-2.3.0
+2.5.0
