@@ -1,8 +1,8 @@
 source 'https://rubygems.org'
-ruby "2.3.0"
+ruby '2.5.0'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.7.1'
+gem 'rails', '4.2.10'
 gem 'pg'
 gem 'uglifier', '>= 1.3.0'
 gem 'therubyracer'
@@ -20,7 +20,7 @@ gem 'puma'
 gem 'devise'
 gem 'aws-sdk'
 
-gem 'activeadmin', git: 'https://github.com/activeadmin/activeadmin'
+gem 'activeadmin', '1.0.0'
 
 gem 'sitemap_generator'
 
@@ -42,7 +42,7 @@ end
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
   gem 'byebug'
-  gem 'factory_girl_rails'
+  gem 'factory_bot_rails'
   gem 'faker'
 end
 
